requirements.txt
django
django-q
redis
polling
python-dotenv
floyd-cli
Pillow